feat: :art: Renamed reducer to not be plural 'reducers'
