+++ src/app/userDashboard/store/selectors.ts
@@ -1,6 +1,6 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store';
 import { TripState } from '../../models/state';
-import { userDashboardFeatureKey } from './reducers';
+import { userDashboardFeatureKey } from './reducer';
 
 export const selectUserDashboardState = createFeatureSelector<TripState>(
   userDashboardFeatureKey
